observability_stack/applications/python-service/requirement.txt
Flask==2.3.3
requests==2.31.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-flask==0.42b0
opentelemetry-instrumentation-requests==0.42b0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
prometheus-client==0.19.0
gunicorn==21.2.0